requirements.txt
attrs
click
ipdb
pymisp
rich
